% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaRMST.dt.R
\name{AdaRMST.dt}
\alias{AdaRMST.dt}
\title{Discrete (Grid) Search for Adaptive Restricted Mean Survival Time}
\usage{
AdaRMST.dt(
  survival_data,
  alpha = 0.05,
  min_time_fix = 0.2,
  max_time_fix = 1,
  c = 0.08/(max_time_fix - min_time_fix)^2,
  L.grid = seq(min_time_fix, max_time_fix, length.out = floor(1.5 *
    nrow(survival_data)^0.25)),
  L.tilde.index = floor((length(L.grid) + 1)/2)
)
}
\arguments{
\item{survival_data}{A data frame containing survival information with columns:
\itemize{
\item \strong{X}: Observed event/censoring time. Default time unit: year.
\item \strong{delta}: Event indicator (1 = event, 0 = censored).
\item \strong{A}: A binary indicator of treatment arm (0 = control, 1 = treatment).
}}

\item{alpha}{A numeric value for the nominal significance level (default is \code{0.05}). A (1 - alpha) asymptotic confidence interval will be returned.}

\item{min_time_fix}{The lower bound of the time range used in the grid search
(default is \code{0.2}).}

\item{max_time_fix}{The upper bound of the time range used in the grid search
(default is \code{1}).}

\item{c}{A numeric penalty coefficient applied to deviations from \code{L.tilde}
(default is \code{0.08 / (max_time_fix - min_time_fix)^2} with year as unit).}

\item{L.grid}{A numeric vector of candidate restriction times (default is
\code{seq(min_time_fix, max_time_fix, length.out = floor(1.5* nrow(survival_data)^0.25))} according to infill aymptotic regime results).}

\item{L.tilde.index}{An integer index indicating which element of \code{L.grid} is
considered \code{L.tilde} (default is the midpoint:
\code{floor((length(L.grid)+1)/2)}).}
}
\value{
A named list with the following components:
\describe{
\item{\strong{K_star}}{The estimated treatment effect (difference in RMST) at \code{L_star}.}
\item{\strong{L_star}}{The selected optimal restriction time from the grid.}
\item{\strong{ci.K}}{Lower and upper bounds of the (1-alpha) confidence interval for \code{K_star}.}
\item{\strong{p}}{A \emph{p}-value associated with the treatment effect measured by RMST difference.}
}
}
\description{
Performs a discrete (grid-based) search for the optimal restriction time (\code{L_star})
in calculating the Restricted Mean Survival Time (RMST) difference between two arms.
The function evaluates a penalized objective function at each time in \code{L.grid}
(filtered by \code{max_followup_time}), selecting the time that maximizes this criterion
as \code{L_star}. It returns the corresponding treatment effect (\code{K_star}),
confidence interval, and a \emph{p}-value.
}
\references{
This function implements the method described in:

Jinghao Sun, Douglas E. Schaubel, Eric J. Tchetgen Tchetgen (2025).
\emph{Beyond Fixed Restriction Time: Adaptive Restricted Mean Survival Time Methods in Clinical Trials}.
}
