% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaRMST.hulc.R
\name{AdaRMST.hulc}
\alias{AdaRMST.hulc}
\title{HulC-Based Confidence Interval for Adaptive RMST}
\usage{
AdaRMST.hulc(
  survival_data,
  alpha = 0.05,
  type = "anti-conservative",
  min_time_fix = 0.2,
  max_time_fix = 1,
  tol = 1e-07,
  maxeval = 70,
  verbose = FALSE
)
}
\arguments{
\item{survival_data}{A data frame containing survival times and event indicators.
Must have:
\itemize{
\item \strong{X}: Observed time (event or censoring). Default time unit: years.
\item \strong{delta}: Event indicator (1 = event, 0 = censored).
\item \strong{A}: Treatment arm indicator (0 = control, 1 = treatment).
}}

\item{alpha}{Numeric value for the nominal significance level (default is \code{0.05}).
Used to decide \code{B} and report (1-alpha) confidence intervals.}

\item{type}{A character string indicating the rule for choosing \code{B}.
Either \code{"anti-conservative"} or \code{"conservative"}.}

\item{min_time_fix}{Lower bound of the time range for RMST estimation
(default is \code{0.2}).}

\item{max_time_fix}{Upper bound of the time range for RMST estimation
(default is \code{1}).}

\item{tol}{A numeric tolerance for optimization (default is \code{1e-7}).}

\item{maxeval}{The maximum number of function evaluations for optimization
(default is \code{70}).}

\item{verbose}{Logical indicating whether to print verbose messages (default is \code{FALSE}).}
}
\value{
A named list with the following elements:
\describe{
\item{\strong{K_star}}{The overall RMST difference estimate from the full dataset.}
\item{\strong{L_star}}{The overall optimal restriction time estimate from the full dataset.}
\item{\strong{ci.K}}{Lower and upper bounds of the (1-alpha) confidence interval for \code{K_star}.}
}
}
\description{
HulC-Based Confidence Interval for Adaptive RMST
}
\references{
This function implements the method described in:

Jinghao Sun, Douglas E. Schaubel, Eric J. Tchetgen Tchetgen (2025).
\emph{Beyond Fixed Restriction Time: Adaptive Restricted Mean Survival Time Methods in Clinical Trials}.
}
\seealso{
\code{\link{AdaRMST.ct}} for point estimation of the adaptive RMST,
\code{\link{choose_B_floor}}, \code{\link{choose_B_ceiling}} for determining \code{B}.
}
