% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{choose_B_ceiling}
\alias{choose_B_ceiling}
\title{Compute the Number of Folds for HulC: Conservative Approach}
\usage{
choose_B_ceiling(alpha = 0.05)
}
\arguments{
\item{alpha}{Numeric. Significance level for determining the number of folds.
Common choices are \code{0.05} for a 95\\% confidence level or \code{0.01} for a 99\\% confidence level.
The default is \code{0.05}.}
}
\value{
Integer. The calculated number of split folds (\code{B}) for the HULC method.
}
\description{
Calculates the number of split folds (\code{B}) required for HulC
method using a conservative rule based on the specified significance level (\code{alpha}).
}
\details{
The conservative version ensures that the number of folds is rounded up to the nearest integer,
providing a higher number of splits which may lead to more robust confidence interval estimates.
}
