% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{choose_B_floor}
\alias{choose_B_floor}
\title{Compute the Number of Folds for HulC: Anti-Conservative Approach}
\usage{
choose_B_floor(alpha = 0.05)
}
\arguments{
\item{alpha}{Numeric. Significance level for determining the number of folds.
Common choices are \code{0.05} for a 95\\% confidence level or \code{0.01} for a 99\\% confidence level.
The default is \code{0.05}.}
}
\value{
Integer. The calculated number of split folds (\code{B}) for the HULC method.
}
\description{
Calculates the number of split folds (\code{B}) required for HulC
method using an anti-conservative rule based on the specified significance level (\code{alpha}).
}
\details{
The anti-conservative version ensures that the number of folds is rounded down to the nearest integer,
providing a lower number of splits which may increase power.
}
